Terminal STDOUT functionality via pipes in place
I used the existing pipe functionality, created a terminal struct and a
TTY struct.  Hooked the terminal class into the scheduler, and after
every schedule, the terminal code reads the STDOUT pipes and updates
screen buffers for any STDOUT written.  If the STDOUT it is writing to
is the "active" TTY then it dups what it wrote to the buffer to the real
screen buffer. (i.e. 0xB8000)